ArticleFacade methods used only by FE API moved to FE API
