+++ packages/framework/src/Model/Article/ArticleFacade.php
@@ -110,64 +110,6 @@ class ArticleFacade
     }
 
     /**
-     * @param int $domainId
-     * @return int
-     */
-    public function getAllVisibleArticlesCountByDomainId(int $domainId): int
-    {
-        return $this->articleRepository->getAllVisibleArticlesCountByDomainId($domainId);
-    }
-
-    /**
-     * @param int $domainId
-     * @param string $placement
-     * @return int
-     */
-    public function getAllVisibleArticlesCountByDomainIdAndPlacement(int $domainId, string $placement): int
-    {
-        return $this->articleRepository->getAllVisibleArticlesCountByDomainIdAndPlacement($domainId, $placement);
-    }
-
-    /**
-     * @param int $domainId
-     * @param int $limit
-     * @param int $offset
-     * @return \Shopsys\FrameworkBundle\Model\Article\Article[]
-     */
-    public function getVisibleArticlesListByDomainId(
-        int $domainId,
-        int $limit,
-        int $offset
-    ): array {
-        return $this->articleRepository->getVisibleListByDomainId(
-            $domainId,
-            $limit,
-            $offset
-        );
-    }
-
-    /**
-     * @param int $domainId
-     * @param string $placement
-     * @param int $limit
-     * @param int $offset
-     * @return \Shopsys\FrameworkBundle\Model\Article\Article[]
-     */
-    public function getVisibleArticlesListByDomainIdAndPlacement(
-        int $domainId,
-        string $placement,
-        int $limit,
-        int $offset
-    ): array {
-        return $this->articleRepository->getVisibleListByDomainIdAndPlacement(
-            $domainId,
-            $placement,
-            $limit,
-            $offset
-        );
-    }
-
-    /**
      * @param \Shopsys\FrameworkBundle\Model\Article\ArticleData $articleData
      * @return \Shopsys\FrameworkBundle\Model\Article\Article
      */
@@ -259,15 +201,5 @@ class ArticleFacade
             t('in footer') => Article::PLACEMENT_FOOTER,
             t('without positioning') => Article::PLACEMENT_NONE,
         ];
-    }
-
-    /**
-     * @param int $domainId
-     * @param string $uuid
-     * @return \Shopsys\FrameworkBundle\Model\Article\Article
-     */
-    public function getVisibleByDomainIdAndUuid(int $domainId, string $uuid): Article
-    {
-        return $this->articleRepository->getVisibleByDomainIdAndUuid($domainId, $uuid);
     }
 }
