+++ packages/framework/src/Model/Article/ArticleRepository.php
@@ -4,7 +4,6 @@ namespace Shopsys\FrameworkBundle\Model\Article;
 
 use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\ORM\QueryBuilder;
-use Shopsys\FrameworkBundle\Model\Article\Exception\ArticleNotFoundException;
 
 class ArticleRepository
 {
@@ -97,42 +96,6 @@ class ArticleRepository
     }
 
     /**
-     * @param int $domainId
-     * @param string $placement
-     * @param int $limit
-     * @param int $offset
-     * @return \Shopsys\FrameworkBundle\Model\Article\Article[]
-     */
-    public function getVisibleListByDomainIdAndPlacement(
-        int $domainId,
-        string $placement,
-        int $limit,
-        int $offset
-    ): array {
-        $queryBuilder = $this->getVisibleArticlesByDomainIdAndPlacementSortedByPositionQueryBuilder($domainId, $placement)
-            ->setFirstResult($offset)
-            ->setMaxResults($limit);
-
-        return $queryBuilder->getQuery()->execute();
-    }
-
-    /**
-     * @param int $domainId
-     * @param string $placement
-     * @return int
-     */
-    public function getAllVisibleArticlesCountByDomainIdAndPlacement(int $domainId, string $placement): int
-    {
-        $queryBuilder = $this->getArticlesByDomainIdQueryBuilder($domainId)
-            ->select('COUNT(a)')
-            ->andWhere('a.hidden = false')
-            ->andWhere('a.placement = :placement')
-            ->setParameter('placement', $placement);
-
-        return (int)$queryBuilder->getQuery()->getSingleScalarResult();
-    }
-
-    /**
      * @param int $articleId
      * @return \Shopsys\FrameworkBundle\Model\Article\Article
      */
@@ -177,41 +140,6 @@ class ArticleRepository
 
     /**
      * @param int $domainId
-     * @return int
-     */
-    public function getAllVisibleArticlesCountByDomainId($domainId): int
-    {
-        $queryBuilder = $this->getArticlesByDomainIdQueryBuilder($domainId)
-            ->select('COUNT(a)')
-            ->andWhere('a.hidden = false');
-
-        return (int)$queryBuilder->getQuery()->getSingleScalarResult();
-    }
-
-    /**
-     * @param int $domainId
-     * @param int $limit
-     * @param int $offset
-     * @return \Shopsys\FrameworkBundle\Model\Article\Article[]
-     */
-    public function getVisibleListByDomainId(
-        int $domainId,
-        int $limit,
-        int $offset
-    ): array {
-        $queryBuilder = $this->getAllVisibleQueryBuilder()
-            ->andWhere('a.domainId = :domainId')
-            ->setParameter('domainId', $domainId)
-            ->orderBy('a.placement')
-            ->addOrderBy('a.position')
-            ->setFirstResult($offset)
-            ->setMaxResults($limit);
-
-        return $queryBuilder->getQuery()->execute();
-    }
-
-    /**
-     * @param int $domainId
      * @return \Shopsys\FrameworkBundle\Model\Article\Article[]
      */
     public function getAllByDomainId($domainId)
@@ -234,26 +162,5 @@ class ArticleRepository
             ->andWhere('a.placement = :placement')->setParameter('placement', $placement)
             ->orderBy('a.position, a.id');
         return $queryBuilder;
-    }
-
-    /**
-     * @param int $domainId
-     * @param string $uuid
-     * @return \Shopsys\FrameworkBundle\Model\Article\Article
-     */
-    public function getVisibleByDomainIdAndUuid(int $domainId, string $uuid): Article
-    {
-        $article = $this->getAllVisibleQueryBuilder()
-            ->andWhere('a.domainId = :domainId')
-            ->setParameter('domainId', $domainId)
-            ->andWhere('a.uuid = :uuid')
-            ->setParameter('uuid', $uuid)
-            ->getQuery()->getOneOrNullResult();
-
-        if ($article === null) {
-            $message = 'Article with UUID \'' . $uuid . '\' not found.';
-            throw new ArticleNotFoundException($message);
-        }
-        return $article;
     }
 }
